[/Blog]: posts: refactoring posts section [splite component to smallest components]
