+++ components/pages/Blog/Posts/PostCardComponent/PostCardComponent.component.tsx
@@ -2,10 +2,8 @@ import s from "./PostCardComponent.component.module.scss";
 
 import Image from "next/image";
 
-
-
 import AnimationWrapper from "@/components/UI/AnimationWrapper/AnimationWrapper.component";
-import { PostCardType } from "@/types/PostType.type";
+import { type PostCardType } from "@/types/PostType.type";
 
 function PostCardComponent({
   title,
