+++ components/pages/Blog/Posts/PostsButton/PostsButton.component.module.scss
@@ -1,0 +1,5 @@
+@import "@/style/index.scss";
+
+.buttonWrapper {
+  @include d-flex;
+}
