+++ components/pages/Blog/Posts/PostsButton/PostsButton.component.tsx
@@ -1,0 +1,12 @@
+import Button from "@/components/UI/Button/Button.component";
+import s from './PostsButton.component.module.scss';
+
+function PostsButton() {
+  return (
+    <section className={s.buttonWrapper}>
+      <Button value="Więcej" type="normal" />
+    </section>
+  );
+}
+
+export default PostsButton;
